end of week 3 commit

+++ Semestru 3/MAP/lab/lab3/lab3/src/service/ServiceUtilizator.java
@@ -23,6 +23,9 @@ public class ServiceUtilizator {
     }
 
     public void deleteUtilizator (long id) {
+        if (this.repo.findOne(id) == null) {
+            throw new ValidationException("User not found");
+        }
         try {
             this.repo.delete(id);
         } catch (ValidationException e) {
